moved the helper functions and changed the format for load_preset
